requirements.txt
# SOP Manager - Image Management Requirements
# Last Updated: December 2024

# Core Supabase SDK
supabase==2.17.0

# HTTP requests for URL testing
requests==2.31.0

# Data handling and processing
pandas==2.1.4
numpy==1.24.3

# Image processing (if needed for future enhancements)
Pillow==10.1.0

# Environment and configuration
python-dotenv==1.0.0

# JSON handling
jsonschema==4.20.0

# Progress bars and utilities
tqdm==4.66.1

# Date and time handling
python-dateutil==2.8.2

# Optional: For advanced image processing
# opencv-python==4.8.1.78

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.11.0
flake8==6.1.0 